+++ src/entities/Player.cpp
@@ -48,11 +48,11 @@ void Player::handle(KeyEvent kEv)
 		r.x = fixtoi(x);
 		r.y = fixtoi(y);
 		
-		GS->DC->add(img[((isSwitchingPolarity / 4) * 2) + polarity], &r, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+		GS->DC->add(img[((isSwitchingPolarity / 4) * 2) + polarity], &r, false, Constants::CamRelation::DEFAULT);
 		
 		if(deathCounter)
 		{
-			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_SHIP_INVINCIBLE_LIGHT, polarity), &r, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_SHIP_INVINCIBLE_LIGHT, polarity), &r, false, Constants::CamRelation::DEFAULT);
 			deathCounter--;
 		}
 		
@@ -114,8 +114,8 @@ void Player::handle(KeyEvent kEv)
 				if(fireRepeat)
 				{
 					// fire 2 bullets if the key is being held
-					Level::bArray->add(x - itofix(img[0][0]) / 4, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, static_cast<int>(Constants::CamRelation::DEFAULT));
-					Level::bArray->add(x + itofix(img[0][0]) / 4, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+					Level::bArray->add(x - itofix(img[0][0]) / 4, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, Constants::CamRelation::DEFAULT);
+					Level::bArray->add(x + itofix(img[0][0]) / 4, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, Constants::CamRelation::DEFAULT);
 					fireDelay = 4;
 					int x1 = x - itofix(img[0][0]) / 4;
 					int x2 = x + itofix(img[0][0]) / 4;
@@ -128,7 +128,7 @@ void Player::handle(KeyEvent kEv)
 				else
 				{
 					// fire 1 bullet
-					Level::bArray->add(x, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+					Level::bArray->add(x, y, 192, itofix(6), static_cast<int>(LUTs::BaseImageId::PLAYER_BULLET_LIGHT), polarity, false, Constants::CamRelation::DEFAULT);
 					fireDelay = 8;
 					fireRepeat = true;
 					for(int i = 0; i < 8; i++)
@@ -152,7 +152,7 @@ void Player::handle(KeyEvent kEv)
 		{
 			r.x = fixtoi(x);
 			r.y = fixtoi(y);
-			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_EXPLOSION_0, deathCounter), &r, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_EXPLOSION_0, deathCounter), &r, false, Constants::CamRelation::DEFAULT);
 			// Death animation
 			// Uses frameskipping as a counter
 			//if(!(GS->skipFrame % 8))
@@ -173,7 +173,7 @@ void Player::handle(KeyEvent kEv)
 		{
 			r.x = fixtoi(x);
 			r.y = fixtoi(y);
-			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_SHIP_LIGHT), &r, false, static_cast<int>(Constants::CamRelation::DEFAULT));
+			GS->DC->add(LUTs::baseImage(LUTs::BaseImageId::PLAYER_SHIP_LIGHT), &r, false, Constants::CamRelation::DEFAULT);
 			y -= itofix(2);
 		}
 		else active = true;
